# Optimize internal `extract!` calls to save on memory allocation